requirements.txt
duckdb>=1.0.0
pyyaml>=6.0
pandas>=2.0.0
scikit-learn>=1.3.0
shap>=0.43.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
